refactor: auth 상태를 전역화하고 헤더 로딩 공백을 제거한다
