+++ src/app/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { Inter, Outfit } from "next/font/google";
+import { AppProviders } from "./providers";
 import "./globals.css";
 
 const inter = Inter({
@@ -25,7 +26,7 @@ export default function RootLayout({
   return (
     <html lang="ko" className={`${inter.variable} ${outfit.variable}`}>
       <body className="bg-slate-50 font-sans text-slate-900 antialiased selection:bg-teal-200 selection:text-teal-900">
-        {children}
+        <AppProviders>{children}</AppProviders>
       </body>
     </html>
   );
